Refactored components and actions
