+++ project/src/store/store.ts
@@ -3,7 +3,7 @@ import { createAPI } from '../services/api/api';
 import { redirect } from './middlewares/redirect';
 import mainPageReducer from './main-page/main-page-reducer';
 import moviePageReducer from './similar-movies/similar-movies-reducer';
-import moviePlayerReducer from './movie/current-movie-reducer';
+import moviePlayerReducer from './movie/movie-reducer';
 import userReducer from './user/user-reducer';
 
 export const rootReducer = combineReducers({mainPageReducer, moviePageReducer, moviePlayerReducer, userReducer});
